split into two scripts
